HSP3dish、中心座標がずれる問題を修正
git-svn-id: http://dev.onionsoft.net/svn/openhsp/trunk@496 ceb3ebad-0e53-d446-8596-a45b2b8d79cd

+++ hsp3dish/hgio.h
@@ -32,7 +32,7 @@ void hgio_line2( float x, float y );
 void hgio_boxf( BMSCR *bm, float x1, float y1, float x2, float y2 );
 void hgio_circle( BMSCR *bm, float x1, float y1, float x2, float y2, int mode );
 void hgio_copy( BMSCR *bm, short xx, short yy, short srcsx, short srcsy, BMSCR *bmsrc, float psx, float psy );
-void hgio_copyrot( BMSCR *bm, short xx, short yy, short srcsx, short srcsy, short ofsx, short ofsy, BMSCR *bmsrc, float psx, float psy, float ang );
+void hgio_copyrot( BMSCR *bm, short xx, short yy, short srcsx, short srcsy, float ofsx, float ofsy, BMSCR *bmsrc, float psx, float psy, float ang );
 void hgio_fillrot( BMSCR *bm, float x, float y, float sx, float sy, float ang );
 
 void hgio_clsmode( int mode, int color, int tex );
